+++ src/Starcounter.XSON/Sessions/Session.cs
@@ -80,13 +80,7 @@ namespace Starcounter {
         /// and no siblings should be serialized.
         /// </summary>
         internal bool enableNamespaces = false;
-
-        /// <summary>
-        /// If set to true values that are bound will not be read several times when generating
-        /// changes from changelog and creating patches.
-        /// </summary>
-        internal bool enableCachedReads = false;
-
+        
         public Session() : this(SessionOptions.Default) {
         }
 
@@ -188,16 +182,11 @@ namespace Starcounter {
                 return;
 
             // Calculating the patch.
-            Byte[] patch;
-            Int32 sizeBytes = jsonPatch_.Generate(
-                PublicViewModel, 
-                true, 
-                CheckOption(SessionOptions.IncludeNamespaces), 
-                out patch);
-
-            if (sizeBytes >= 0) {
+            string patch = jsonPatch_.Generate(PublicViewModel, true, CheckOption(SessionOptions.IncludeNamespaces));
+            
+            if (!string.IsNullOrEmpty(patch)) {
                 // Sending the patch bytes to the client.
-                ActiveWebSocket.Send(patch, sizeBytes, true);
+                ActiveWebSocket.Send(patch);
             }
         }
 
@@ -617,12 +606,7 @@ namespace Starcounter {
                 return true;
             }
         }
-        
-        /// <summary>
-        /// 
-        /// </summary>
-        public bool HaveAddedOrRemovedObjects { get; set; }
-        
+      
         internal TransactionHandle RegisterTransaction(TransactionHandle handle) {
             TransactionRef tref = null;